requirements.txt
# Core dependencies
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# CLI
click==8.1.7
rich==13.9.4

# Data processing
pandas==2.2.3
openpyxl==3.1.5

# HTTP clients
requests==2.32.3
aiohttp==3.11.10
tenacity==9.0.0

# Web scraping
beautifulsoup4==4.12.3
lxml==5.3.0

# API clients (DataforSEO requires base64 auth)
base64io==1.0.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0

# Code quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0
types-requests==2.32.0.20241016

# Logging
loguru==0.7.2